+++ src/rust/lib/shortcuts/Cargo.toml
@@ -1,12 +1,28 @@
 [package]
 name    = "enso-shortcuts"
 version = "0.1.0"
-authors = ["Enso Team <contact@luna-lang.org>"]
+authors = ["Enso Team <contact@enso.org>"]
 edition = "2018"
 
+
 [lib]
+crate-type = ["cdylib", "rlib"]
+
 
 [dependencies]
-enso-automata = { path = "../automata" }
-enso-prelude  = { path = "../enso-prelude" }
-enso-frp      = { path = "../frp" }
+enso-frp          = { path = "../frp"               }
+enso-automata     = { path = "../automata"          }
+enso-prelude      = { path = "../enso-prelude"      }
+ensogl-system-web = { path = "../../lib/system/web" }
+logger            = { path = "../../lib/logger"     }
+
+wasm-bindgen = { version = "=0.2.58" , features = ["nightly","serde-serialize"] }
+nalgebra     = { version = "0.21.1"  , features = ["serde-serialize"]           }
+serde_json   = { version = "1.0"                                                }
+serde        = { version = "1.0"     , features = ["derive"]                    }
+js-sys       = { version = "0.3.28"                                             }
+
+
+[dependencies.web-sys]
+version = "0.3.4"
+features = []
